Housekeeping: Add feature flag / POC / Bedrock checkboxes to PR template
Another reason for not adding automated tests (yet) is that the code doesn't go live yet, but instead is hidden behind a feature flag. A future PR will add tests and make the feature permanent by removing the feature flag.
Code may also be a proof of concept, to be rewritten (or even removed) later once the feasibility has been checked. This may be combined with a feature flag, too.
The Bedrock checkbox may only be applicable to some PRs (e.g. for the resource pack), but like with LTS and Docker support, it's a good reminder that this has to be considered.